requirements.txt
numpy==1.24.3
redis==5.0.1
websockets==12.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
tensorflow==2.15.0
pyaudio==0.2.14
dlib
face-recognition==1.3.0

